runtime.package refactored to not use FP
It seems that State was a bad abstraction to begin with.
The problem with State is that A => State[A] functions may
modify said state when they return. Hence I had to come up
with the `bracket` function to prevent such modifications.
If we don't use the State abstraction to begin with, such
modification problems do not arise. This commit is mostly
about passing the context implicitly rather than with a State.